added input versioning, fixed a cancellation race condition on the game task

+++ lawliet/lawliet-core/src/ability/anonymous_announcement.rs
@@ -16,7 +16,7 @@ impl AbilityInterface for AnonymousAnnouncement {
         ctx: &mut ActionContext,
         _: &crate::action::ActionActor,
         _: AbilityKey,
-        _: u8,
+        _: u64,
         _mutate: bool,
     ) -> super::AbilityResult {
         cmd_world_event(
