+++ lawliet/lawliet-core/src/ability/false_anonymous_contact.rs
@@ -23,7 +23,7 @@ impl AbilityInterface for FalseAnonymousContact {
         ctx: &mut crate::action::ActionContext,
         actor: &crate::action::ActionActor,
         _ability: crate::AbilityKey,
-        version: u8,
+        version: u64,
         mutate: bool,
     ) -> super::AbilityResult {
         actor.player_only()?;
