+++ lawliet/lawliet-core/src/ability/force_invite.rs
@@ -20,7 +20,7 @@ impl AbilityInterface for ForceInvite {
         ctx: &mut crate::action::ActionContext,
         actor: &crate::action::ActionActor,
         _ability: crate::AbilityKey,
-        version: u8,
+        version: u64,
         mutate: bool,
     ) -> super::AbilityResult {
         actor.org_only()?;
