+++ lawliet/lawliet-core/src/ability/true_name_reveal.rs
@@ -24,7 +24,7 @@ impl AbilityInterface for TrueNameReveal {
         ctx: &mut crate::action::ActionContext,
         actor: &crate::action::ActionActor,
         _ability: crate::AbilityKey,
-        _version: u8,
+        _version: u64,
         _mutate: bool,
     ) -> super::AbilityResult {
         actor.player_only()?;
