+++ lawliet/lawliet-core/src/ability/unlawful_arrest.rs
@@ -21,7 +21,7 @@ impl AbilityInterface for UnlawfulArrest {
         ctx: &mut lawliet_types::action::ActionContext,
         _actor: &lawliet_types::action::ActionActor,
         ability: lawliet_types::common::AbilityKey,
-        version: u8,
+        version: u64,
         mutate: bool,
     ) -> super::AbilityResult {
         // You can only arrest a player; the arrester need not be one.
